removed ./ultils.py which contains only one function
