+++ tests/ci/test_import_graph.py
@@ -262,14 +262,14 @@ class TestResolveTestTargets:
         fake_pkg(
             {
                 "probpipe/__init__.py": "",
-                "probpipe/_utils.py": "v = 1\n",
-                "tests/test_utils.py": "",
+                "probpipe/_array_utils.py": "v = 1\n",
+                "tests/test_array_utils.py": "",
                 "tests/test_coverage_gaps.py": "",
                 "tests/core/test_thing.py": "",  # subpackage test — must NOT be selected
             }
         )
-        assert set(resolve_test_targets(["probpipe/_utils.py"])) == {
-            "tests/test_utils.py",
+        assert set(resolve_test_targets(["probpipe/_array_utils.py"])) == {
+            "tests/test_array_utils.py",
             "tests/test_coverage_gaps.py",
         }
